+++ src/main/java/com/example/ehe_server/service/intf/auth/JwtTokenGeneratorInterface.java
@@ -1,6 +1,6 @@
 package com.example.ehe_server.service.intf.auth;
 
 public interface JwtTokenGeneratorInterface {
-    String generateAccessToken(Long userId, String role);
-    String generateRefreshToken(Long userId, String role);
+    String generateAccessToken(Integer userId, String role);
+    String generateRefreshToken(Integer userId, String role);
 }
